+++ git/refs/remote.py
@@ -1,16 +1,18 @@
-from head import Head
-from git.util import join_path
-from gitdb.util import join
-
 import os
-
+from headref import Head
+from git.util import (
+						join, 
+						join_path
+						)
 
 __all__ = ["RemoteReference"]
 
 	
 class RemoteReference(Head):
 	"""Represents a reference pointing to a remote head."""
-	_common_path_default = "refs/remotes"
+	__slots__ = tuple()
+	
+	_common_path_default = Head._remote_common_path_default
 	
 	
 	@classmethod
@@ -22,24 +24,11 @@ class RemoteReference(Head):
 		# END handle remote constraint
 		return super(RemoteReference, cls).iter_items(repo, common_path)
 	
-	@property
-	def remote_name(self):
-		"""
-		:return:
-			Name of the remote we are a reference of, such as 'origin' for a reference
-			named 'origin/master'"""
-		tokens = self.path.split('/')
-		# /refs/remotes/<remote name>/<branch_name>
-		return tokens[2]
-		
-	@property
-	def remote_head(self):
-		""":return: Name of the remote head itself, i.e. master.
-		:note: The returned name is usually not qualified enough to uniquely identify
-			a branch"""
-		tokens = self.path.split('/')
-		return '/'.join(tokens[3:])
-		
+	@classmethod
+	def create(cls, *args, **kwargs):
+		"""Used to disable this method"""
+		raise TypeError("Cannot explicitly create remote references")
+	
 	@classmethod
 	def delete(cls, repo, *refs, **kwargs):
 		"""Delete the given remote references.
@@ -56,8 +45,3 @@ class RemoteReference(Head):
 			except OSError:
 				pass
 		# END for each ref
-		
-	@classmethod
-	def create(cls, *args, **kwargs):
-		"""Used to disable this method"""
-		raise TypeError("Cannot explicitly create remote references")